requirements.txt
appdirs==1.4.4
argcomplete==1.12.3
argon2-cffi==21.1.0
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
cryptography==35.0.0
cssselect==1.1.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
entrypoints==0.3
fake-useragent==0.1.11
feedparser==6.0.8
Flask==2.0.2
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.8.1
importlib-resources==5.4.0
ipykernel==6.5.0
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.2
joblib==1.1.0
jsonschema==4.2.1
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.8
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.5
numpy==1.21.3
packaging==21.2
pandas==1.3.4
pandocfilters==1.5.0
parse==1.19.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prettytable==2.2.1
prometheus-client==0.12.0
prompt-toolkit==3.0.22
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pycparser==2.20
pyecharts==1.9.0
pyee==8.2.2
Pygments==2.10.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyppeteer==0.2.6
pyquery==1.4.3
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
pyzmq==22.3.0
qtconsole==5.2.0
QtPy==1.11.2
requests==2.26.0
requests-html==0.10.0
scikit-learn==1.0
scipy==1.7.1
Send2Trash==1.8.0
sgmllib3k==1.0.0
simplejson==3.17.5
six==1.16.0
sklearn==0.0
soupsieve==2.2.1
SQLAlchemy==1.4.26
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing-extensions==3.10.0.2
urllib3==1.26.7
w3lib==1.22.0
wcwidth==0.2.5
webencodings==0.5.1
websockets==9.1
Werkzeug==2.0.2
widgetsnbextension==3.5.2
yahoo-fin==0.8.9.1
zipp==3.6.0
